MVP Model Done
Image Dataset is too large for Github. If you want access to the image dataset, message me at https://evidencen.com/contact-me/

Maybe, in the future i will figure out how to effectively compress the images to a small file size and be able to upload it here on Github.